backend/requirements.txt
# Data Processing & Visualization
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.28
matplotlib
seaborn

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.15.0
xgboost>=2.0.0

# Financial Time Series
arch>=6.0.0

# AI Agents - Core Dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0

# AI Agents - Database
pymongo>=4.0.0

# AI Agents - HTTP & APIs
requests>=2.31.0

# AI Agents - NLP & Transformers
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# AI Agents - LLM APIs
google-generativeai>=0.3.0
groq>=0.4.0

# Quantitative Strategy Library - LangChain/LangGraph
langchain>=0.1.0
langchain-google-genai>=0.0.5
langgraph>=0.0.20
scipy>=1.10.0

# Optional: For local LLM inference
huggingface-hub>=0.19.0

# ============================================
# RAG Trading Mentor - Additional Dependencies
# ============================================

# Vector Database
chromadb>=0.4.0

# Embedding Models (Free/Open Source)
sentence-transformers>=2.2.0

# Logging (Optional)
coloredlogs>=15.0

# Database Fallback
tinydb>=4.8.0

# Web API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

